+++ __tests__/unit/backend-proxy.test.js
@@ -14,6 +14,7 @@ describe('Backend Proxy', () => {
 	};
 
 	let mockRequest;
+	let mockResponse;
 	let next;
 	let proxyRequest;
 	let backendResponse;
@@ -31,6 +32,8 @@ describe('Backend Proxy', () => {
 				'some-header': 'header-value'
 			}
 		};
+
+		mockResponse = {};
 
 		next = jest.fn();
 
@@ -64,7 +67,7 @@ describe('Backend Proxy', () => {
 		mockRequest[baseOptions.key] = true;
 
 		// When
-		middleware(mockRequest, undefined, next);
+		middleware(mockRequest, mockResponse, next);
 
 		// Then
 		expect(next).toHaveBeenCalledTimes(1);
@@ -77,9 +80,13 @@ describe('Backend Proxy', () => {
 		mockRequest.pipe.mockReturnValueOnce({
 			on: jest.fn()
 		});
+		mockRequest.headers = {
+			...mockRequest.headers,
+			host: 'original.host:8080'
+		};
 
 		// When
-		middleware(mockRequest, undefined, next);
+		middleware(mockRequest, mockResponse, next);
 
 		// Then
 		expect(mockRequest.pipe).toHaveBeenCalledTimes(1);
@@ -88,7 +95,11 @@ describe('Backend Proxy', () => {
 			hostname: 'backend.local',
 			path: mockRequest.url,
 			method: mockRequest.method,
-			headers: mockRequest.headers
+			headers: {
+				...mockRequest.headers,
+				host: 'backend.local',
+				'X-Orig-Host': 'original.host:8080'
+			}
 		}));
 		// We haven't simulated an error or a response so next should not have been called at this point
 		expect(next).not.toHaveBeenCalled();
@@ -106,7 +117,7 @@ describe('Backend Proxy', () => {
 		});
 
 		// When
-		middleware(mockRequest, undefined, next);
+		middleware(mockRequest, mockResponse, next);
 
 		// Then
 		expect(request).toHaveBeenCalledWith(expect.objectContaining({
@@ -119,41 +130,11 @@ describe('Backend Proxy', () => {
 		expect(next).not.toHaveBeenCalled();
 	});
 
-	test(`pipes the request to the backend with the host header changed when changeHost is true`, () => {
-		// Given
-		const middleware = backendProxy({
-			...baseOptions,
-			usePath: false,
-			backend: 'http://backend.local/sub/path',
-			changeHost: true
-		});
-		mockRequest.headers = {
-			host: 'original.host:8080'
-		};
-
-		// When
-		middleware(mockRequest, undefined, next);
-
-		// Then
-		expect(request).toHaveBeenCalledWith(expect.objectContaining({
-			hostname: 'backend.local',
-			path: '/sub/path',
-			method: mockRequest.method,
-			headers: {
-				...mockRequest.headers,
-				host: 'backend.local',
-				'X-Orig-Host': 'original.host:8080'
-			}
-		}));
-		// We haven't simulated an error or a response so next should not have been called at this point
-		expect(next).not.toHaveBeenCalled();
-	});
-
 	describe('backend response', () => {
 		test('passes an error on to next', () => {
 			// Given
 			const middleware = backendProxy(baseOptions);
-			middleware(mockRequest, undefined, next);
+			middleware(mockRequest, mockResponse, next);
 			const error = {code: 'some error'};
 
 			// When
@@ -217,8 +198,9 @@ describe('Backend Proxy', () => {
 				field4: 'value4'
 			};
 			const parts = JSON.stringify(backendBody).match(/.{1,2}/g);
+			backendResponse.statusCode = 200;
 
-			middleware(mockRequest, undefined, next);
+			middleware(mockRequest, mockResponse, next);
 			proxyRequest.emit('response', backendResponse);
 
 			// When
@@ -226,6 +208,7 @@ describe('Backend Proxy', () => {
 			backendResponse.emit('end');
 
 			// Then
+			expect(mockResponse.statusCode).toEqual(200);
 			expect(mockRequest[baseOptions.key]).toEqual(backendBody);
 			expect(next).toHaveBeenCalledTimes(1);
 		});
@@ -243,7 +226,7 @@ describe('Backend Proxy', () => {
 			};
 			const parts = JSON.stringify(backendBody).match(/.{1,2}/g);
 
-			middleware(mockRequest, undefined, next);
+			middleware(mockRequest, mockResponse, next);
 			proxyRequest.emit('response', backendResponse);
 
 			// When
@@ -260,7 +243,7 @@ describe('Backend Proxy', () => {
 			const middleware = backendProxy(baseOptions);
 			const parts = ['definitely', 'not', '{};; json'];
 
-			middleware(mockRequest, undefined, next);
+			middleware(mockRequest, mockResponse, next);
 			proxyRequest.emit('response', backendResponse);
 
 			// When
@@ -277,7 +260,7 @@ describe('Backend Proxy', () => {
 		test('catches an error reading the backend stream', () => {
 			// Given
 			const middleware = backendProxy(baseOptions);
-			middleware(mockRequest, undefined, next);
+			middleware(mockRequest, mockResponse, next);
 			proxyRequest.emit('response', backendResponse);
 
 			// When
@@ -290,7 +273,7 @@ describe('Backend Proxy', () => {
 		test('catches the backend stream being aborted', () => {
 			// Given
 			const middleware = backendProxy(baseOptions);
-			middleware(mockRequest, undefined, next);
+			middleware(mockRequest, mockResponse, next);
 			proxyRequest.emit('response', backendResponse);
 
 			// When
@@ -303,7 +286,7 @@ describe('Backend Proxy', () => {
 		test('catches the backend stream being closed', () => {
 			// Given
 			const middleware = backendProxy(baseOptions);
-			middleware(mockRequest, undefined, next);
+			middleware(mockRequest, mockResponse, next);
 			proxyRequest.emit('response', backendResponse);
 
 			// When
@@ -317,7 +300,7 @@ describe('Backend Proxy', () => {
 			test('does not call next when the stream is then closed', () => {
 				// Given
 				const middleware = backendProxy(baseOptions);
-				middleware(mockRequest, undefined, next);
+				middleware(mockRequest, mockResponse, next);
 				proxyRequest.emit('response', backendResponse);
 
 				// When
@@ -330,30 +313,17 @@ describe('Backend Proxy', () => {
 			});
 		});
 
-		describe('when interceptErrors is on', () => {
+		describe('when the backend responds with an error', () => {
 			let middleware;
 
 			beforeEach(() => {
 				middleware = backendProxy({
-					...baseOptions,
-					interceptErrors: true
+					...baseOptions
 				});
 
 				middleware(mockRequest, response, next);
 
 				backendResponse.headers = {};
-			});
-
-			test('does not intercept a 200', () => {
-				// When
-				proxyRequest.emit('response', backendResponse);
-
-				// Then
-				expect(backendResponse.pipe).toHaveBeenCalledTimes(1);
-				expect(backendResponse.pipe).toHaveBeenCalledWith(response);
-				expect(response.header).toHaveBeenCalledWith(backendResponse.headers);
-				expect(response.statusCode).toBe(backendResponse.statusCode);
-				expect(next).not.toHaveBeenCalled();
 			});
 
 			test.each(
@@ -371,8 +341,7 @@ describe('Backend Proxy', () => {
 				expect(response.statusCode).toBe(statusCode);
 				expect(next).toHaveBeenCalledTimes(1);
 				expect(next).toHaveBeenCalledWith({
-					statusCode: statusCode,
-					backendResponse: backendResponse
+					statusCode: statusCode
 				});
 			});
 		});